chore: change name of cargo domain to delivery
