+++ src/domain/delivery/application/use-cases/delivery-driver/get-delivery-driver.ts
@@ -1,5 +1,5 @@
 import { Either, left, right } from '@/core/either'
-import { DeliveryDriver } from '@/domain/cargo/enterprise/entities/delivery-driver'
+import { DeliveryDriver } from '@/domain/delivery/enterprise/entities/delivery-driver'
 import { DeliveryDriverRepository } from '../../repositories/delivery-driver-repository'
 import { ResourceNotFoundError } from '@/core/errors/errors/resource-not-found-error'
