+++ src/domain/delivery/application/use-cases/errors/invalid-order-status-error.ts
@@ -1,5 +1,5 @@
 import { UseCaseError } from '@/core/errors/use-case-error'
-import { OrderStatus } from '@/domain/cargo/enterprise/entities/order'
+import { OrderStatus } from '@/domain/delivery/enterprise/entities/order'
 
 export class InvalidOrderStatusError extends Error implements UseCaseError {
   constructor(
