+++ src/domain/delivery/application/use-cases/order/mark-order-as-available.spec.ts
@@ -1,24 +1,22 @@
 import { InMemoryOrdersRepository } from 'test/respositories/in-memory-orders-repository'
 import { makeOrder } from 'test/factories/make-order'
+import { MarkOrderAsAvailableUseCase } from './mark-order-as-available'
 import { UnauthorizedError } from '@/core/errors/errors/unauthorized-error'
-import { OrderStatus } from '@/domain/cargo/enterprise/entities/order'
+import { OrderStatus } from '@/domain/delivery/enterprise/entities/order'
 import { InvalidOrderStatusError } from '../errors/invalid-order-status-error'
-import { ReturnOrderUseCase } from './return-order'
 
 let inMemoryOrdersRepository: InMemoryOrdersRepository
-let sut: ReturnOrderUseCase
+let sut: MarkOrderAsAvailableUseCase
 
-describe('Return Order', () => {
+describe('Mark Order as Available', () => {
   beforeEach(() => {
     inMemoryOrdersRepository = new InMemoryOrdersRepository()
 
-    sut = new ReturnOrderUseCase(inMemoryOrdersRepository)
+    sut = new MarkOrderAsAvailableUseCase(inMemoryOrdersRepository)
   })
 
-  it('should be able return an order', async () => {
-    const order = makeOrder({
-      status: OrderStatus.PICKED_UP,
-    })
+  it('should be able to mark an order as available', async () => {
+    const order = makeOrder()
 
     inMemoryOrdersRepository.items.push(order)
 
@@ -27,13 +25,11 @@ describe('Return Order', () => {
       role: 'ADMIN',
     })
 
-    console.log(result)
-
     expect(result.isRight()).toBe(true)
     expect(result.value?.order.id).toEqual(order.id)
   })
 
-  it('should not authorized to return an order if the user is not admin', async () => {
+  it('should not authorized to mark an order as available if the user is not admin', async () => {
     const order = makeOrder()
 
     inMemoryOrdersRepository.items.push(order)
@@ -47,7 +43,7 @@ describe('Return Order', () => {
     expect(result.value).toBeInstanceOf(UnauthorizedError)
   })
 
-  it('should not able to return an order if the status order is not "PICKED_UP"', async () => {
+  it('should not able to mark an order as available if the status order is not "added"', async () => {
     const order = makeOrder({
       status: OrderStatus.DELIVERED,
     })
