+++ src/domain/delivery/application/use-cases/order/mark-order-as-available.ts
@@ -1,36 +1,30 @@
 import { Either, left, right } from '@/core/either'
 import { OrdersRepository } from '../../repositories/orders-repository'
 import { ResourceNotFoundError } from '@/core/errors/errors/resource-not-found-error'
-import { UserRole } from '@/domain/cargo/enterprise/entities/user-role'
-import { Order, OrderStatus } from '@/domain/cargo/enterprise/entities/order'
+import { UserRole } from '@/domain/delivery/enterprise/entities/user-role'
+import { Order, OrderStatus } from '@/domain/delivery/enterprise/entities/order'
 import { InvalidOrderStatusError } from '../errors/invalid-order-status-error'
 import { UnauthorizedError } from '@/core/errors/errors/unauthorized-error'
-import { DeliveryDriverRepository } from '../../repositories/delivery-driver-repository'
 
-interface PickupOrderUseCaseRequest {
+interface MarkOrderAsAvailableUseCaseRequest {
   orderId: string
-  deliveryDriverId: string
   role: string
 }
 
-type PickupOrderUseCaseResponse = Either<
+type MarkOrderAsAvailableUseCaseResponse = Either<
   UnauthorizedError | ResourceNotFoundError | InvalidOrderStatusError,
   {
     order: Order
   }
 >
 
-export class PickupOrderUseCase {
-  constructor(
-    private ordersRepository: OrdersRepository,
-    private deliveryDriversRepository: DeliveryDriverRepository,
-  ) {}
+export class MarkOrderAsAvailableUseCase {
+  constructor(private ordersRepository: OrdersRepository) {}
 
   async execute({
     orderId,
-    deliveryDriverId,
     role,
-  }: PickupOrderUseCaseRequest): Promise<PickupOrderUseCaseResponse> {
+  }: MarkOrderAsAvailableUseCaseRequest): Promise<MarkOrderAsAvailableUseCaseResponse> {
     const isValidRole = Object.values(UserRole).includes(role as UserRole)
 
     if (!isValidRole) {
@@ -49,20 +43,11 @@ export class PickupOrderUseCase {
       return left(new ResourceNotFoundError())
     }
 
-    const deliveryDriver =
-      await this.deliveryDriversRepository.findById(deliveryDriverId)
-
-    if (!deliveryDriver) {
-      return left(new ResourceNotFoundError())
+    if (order.status !== OrderStatus.ADDED) {
+      return left(new InvalidOrderStatusError(order.status, OrderStatus.ADDED))
     }
 
-    if (order.status !== OrderStatus.AVAILABLE) {
-      return left(
-        new InvalidOrderStatusError(order.status, OrderStatus.AVAILABLE),
-      )
-    }
-
-    order.status = OrderStatus.PICKED_UP
+    order.status = OrderStatus.AVAILABLE
 
     await this.ordersRepository.save(order)
