+++ src/domain/delivery/application/use-cases/order/pickup-order.spec.ts
@@ -1,7 +1,7 @@
 import { InMemoryOrdersRepository } from 'test/respositories/in-memory-orders-repository'
 import { makeOrder } from 'test/factories/make-order'
 import { UnauthorizedError } from '@/core/errors/errors/unauthorized-error'
-import { OrderStatus } from '@/domain/cargo/enterprise/entities/order'
+import { OrderStatus } from '@/domain/delivery/enterprise/entities/order'
 import { InvalidOrderStatusError } from '../errors/invalid-order-status-error'
 import { PickupOrderUseCase } from './pickup-order'
 import { InMemoryDeliveryDriverRepository } from 'test/respositories/in-memory-delivery-driver-repository'
