+++ src/domain/delivery/application/use-cases/recipient/add-recipient.ts
@@ -1,9 +1,9 @@
 import { Either, left, right } from '@/core/either'
 import { RecipientsRepository } from '../../repositories/recipients-repository'
-import { Recipient } from '@/domain/cargo/enterprise/entities/recipient'
+import { Recipient } from '@/domain/delivery/enterprise/entities/recipient'
 import { ResourceNotFoundError } from '@/core/errors/errors/resource-not-found-error'
 import { NotAllowedError } from '@/core/errors/errors/not-allowed-error'
-import { UserRole } from '@/domain/cargo/enterprise/entities/user-role'
+import { UserRole } from '@/domain/delivery/enterprise/entities/user-role'
 
 interface AddRecipientUseCaseRequest {
   name: string
