+++ src/domain/delivery/application/use-cases/recipient/edit-recipient.ts
@@ -1,8 +1,8 @@
 import { Either, left, right } from '@/core/either'
 import { ResourceNotFoundError } from '@/core/errors/errors/resource-not-found-error'
-import { Recipient } from '@/domain/cargo/enterprise/entities/recipient'
+import { Recipient } from '@/domain/delivery/enterprise/entities/recipient'
 import { RecipientsRepository } from '../../repositories/recipients-repository'
-import { UserRole } from '@/domain/cargo/enterprise/entities/user-role'
+import { UserRole } from '@/domain/delivery/enterprise/entities/user-role'
 import { NotAllowedError } from '@/core/errors/errors/not-allowed-error'
 
 interface EditRecipientUseCaseRequest {
