+++ src/domain/delivery/enterprise/entities/administrator.ts
