+++ test/factories/make-order.ts
@@ -1,5 +1,5 @@
 import { UniqueEntityID } from '@/core/entities/unique-entity-id'
-import { Order, OrderProps } from '@/domain/cargo/enterprise/entities/order'
+import { Order, OrderProps } from '@/domain/delivery/enterprise/entities/order'
 
 export function makeOrder(
   override: Partial<OrderProps> = {},
